# Cummax and Cummin doc update and performance benchmark