… aten/src/ATen/native/cuda/ReduceOpsKernel.cu
… aten/src/ATen/native/native_functions.yaml
… aten/src/THC/THCTensorMathReduce.cuh
… test/test_torch.py
… tools/autograd/derivatives.yaml
… torch/_torch_docs.py