@kostmo 2020-01-23
## :pill: CircleCI build failures summary and remediations
As of commit a7302512:


**None of the build failures appear to be your fault**.

* **1/1** recognized as flaky :snowflake:
    * Re-run these jobs?


## Detailed failure analysis

One may explore the probable reasons each build failed interactively [on the Dr. CI website](https://dr.pytorch.org/commit-details.html?sha1=a7302512f8add9ad6e5894664542a36e855777d0).

### :snowflake: 1 failure recognized as flaky
The following build failures have been detected as flaky and **may not be your fault**:
#### [![See CircleCI build](https://avatars0.githubusercontent.com/ml/7?s=12)](https://circleci.com/gh/pytorch/pytorch/4323876) pytorch_linux_xenial_cuda10_1_cudnn7_py3_gcc7_build (1/1)
**Step:** "Build" ([full log](https://dr.pytorch.org/api/view-log-full?build_id=89859271) | [pattern match details](https://dr.pytorch.org/build-details.html?build_id=89859271)) :snowflake:
<details>
<summary>
<code>Jan 23 17:08:05 E: Failed to fetch https://developer.download.nvidia.com/compute/machine-learning/repos/ubuntu1604/x86_64/./libnccl2_2.5.6-1+cuda10.1_amd64.deb  Operation too slow. Less than 10 bytes/sec transferred the last 120 seconds</code>
</summary>

```
Jan 23 17:04:28   libssh2            pkgs/main/linux-64::libssh2-1.8.2-h1ba5d50_0 
Jan 23 17:04:28   rhash              pkgs/main/linux-64::rhash-1.3.8-h1ba5d50_0 
Jan 23 17:04:28  
Jan 23 17:04:28  
Jan 23 17:04:28 Preparing transaction: ...working... done 
Jan 23 17:04:29 Verifying transaction: ...working... done 
Jan 23 17:04:29 Executing transaction: ...working... done 
Jan 23 17:04:29 + [[ pytorch-linux-xenial-cuda10.1-cudnn7-py3-gcc7-build == *-xenial-cuda10.1-* ]] 
Jan 23 17:04:29 + sudo apt-get -qq update 
Jan 23 17:04:33 + sudo apt-get -qq install --allow-downgrades --allow-change-held-packages libnccl-dev=2.5.6-1+cuda10.1 libnccl2=2.5.6-1+cuda10.1 
Jan 23 17:08:05 E: Failed to fetch https://developer.download.nvidia.com/compute/machine-learning/repos/ubuntu1604/x86_64/./libnccl2_2.5.6-1+cuda10.1_amd64.deb  Operation too slow. Less than 10 bytes/sec transferred the last 120 seconds 
Jan 23 17:08:05  
Jan 23 17:08:05 E: Unable to fetch some archives, maybe run apt-get update or try with --fix-missing? 
Jan 23 17:08:05 + cleanup 
Jan 23 17:08:05 + retcode=100 
Jan 23 17:08:05 + set +x 
Jan 23 17:08:05 =================== sccache compilation log =================== 
Jan 23 17:08:05 =========== If your build fails, please take a look at the log above for possible reasons =========== 
Jan 23 17:08:05 Compile requests                 0 
Jan 23 17:08:05 Compile requests executed        0 
Jan 23 17:08:05 Cache hits                       0 
```

</details>



---

<details><summary>This comment was automatically generated by <a href="https://github.com/kostmo/circleci-failure-tracker/tree/master/docs/from-pull-request-comment">Dr. CI</a> (expand for details).</summary>Follow <a href="https://dr.pytorch.org/admin/comments-opt-out.html">this link to opt-out</a> of these comments for your Pull Requests.

Please report bugs/suggestions on the <a href="https://github.com/kostmo/circleci-failure-tracker/issues">GitHub issue tracker</a>.
</details>