@ngimel 2020-01-28
Where are the benchmarking results? I don't see how cummin/cummax can be faster than min/max unless there's something royally wrong with our min/max implementation, because cummin/cummax have output of the same size as input and are doing twice as many memory reads/writes as min/max, even ignoring that parallelizing scan operations is a more challenging problems than reduction generally.