@anjali411 2020-02-19
oh my bad! will make a PR

> @anjali411 you forgot to update doc for cummin:
> 
> ```
> Returns a namedtuple (values, indices) where values is the cumulative maximum 
> ```
> 
> to
> 
> ```
> Returns a namedtuple (values, indices) where values is the cumulative minimum
> ```